PNBD Dyncov: Improved optimx defaults (higher `itnmax`) (#279)
Significantly increase the default number of optimizer iterations for dyncov models (`itnmax`) from 3000 to 50000. The previous iteration limit is rather low for NelderMead and in practice was often reached. This likely mislead users to believe that the optimization converged when in reality it only stopped because the max num iterations had been reached.